+++ resources/js/admin/components/team/Index.vue
@@ -68,12 +68,13 @@
       },
 
       destroy(id) {
-        let uri = `/api/team/destroy/${id}`;
-        this.axios.delete(uri).then(response => {
-          // this.team.splice(this.team.indexOf(id), 1);
-          this.fetch();
-          this.$notify({type: 'success', text: 'Eintrag gelöscht'});
-        });
+        if(confirm('Bitte löschen bestätigen!')) {
+          let uri = `/api/team/destroy/${id}`;
+          this.axios.delete(uri).then(response => {
+            this.fetch();
+            this.$notify({type: 'success', text: 'Eintrag gelöscht'});
+          });
+        }
       },
       
       clone(id) {
